Moved Test Category Icon to Shared Components
