+++ src/components/test-category-icon/__tests__/test-category-icon.spec.ts
