+++ src/components/test-category-icon/test-category-icon.scss
