+++ src/components/test-category-icon/test-category-icon.ts
