+++ src/pages/journal/__mocks__/journal.module.mock.ts
@@ -6,7 +6,7 @@ import { TimeComponent } from '../components/time/time';
 import { CandidateComponent } from '../components/candidate/candidate';
 import { TestOutcomeComponent } from '../components/test-outcome/test-outcome';
 import { TestCategoryComponent } from '../components/test-category/test-category';
-import { TestCategoryIconComponent } from '../components/test-category-icon/test-category-icon';
+import { TestCategoryIconComponent } from '../../../components/test-category-icon/test-category-icon';
 import { VehicleDetailsComponent } from '../components/vehicle-details/vehicle-details';
 
 @NgModule({
